+++ example/lib/builder_screen.dart
@@ -1,3 +1,4 @@
+import 'package:example/util/screen.dart';
 import 'package:flutter/material.dart';
 import 'package:responsive_ui/responsive_ux.dart';
 
@@ -6,34 +7,33 @@ class BuilderScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Scaffold(
-      body: SafeArea(
-        child: WindowSizeBuilder(
-          builder: (width, height, child) {
-            String w = "";
-            String h = "";
+    return Screen(
+      title: "Builder Screen",
+      child: WindowSizeBuilder(
+        builder: (width, height, child) {
+          String w = "";
+          String h = "";
 
-            switch (width) {
-              case WindowSizes.compact:
-                w = "Compact";
-              case WindowSizes.medium:
-                w = "Medium";
-              case WindowSizes.expanded:
-                w = "Expanded";
-            }
+          switch (width) {
+            case WindowSizes.compact:
+              w = "Compact";
+            case WindowSizes.medium:
+              w = "Medium";
+            case WindowSizes.expanded:
+              w = "Expanded";
+          }
 
-            switch (height) {
-              case WindowSizes.compact:
-                h = "Compact";
-              case WindowSizes.medium:
-                h = "Medium";
-              case WindowSizes.expanded:
-                h = "Expanded";
-            }
+          switch (height) {
+            case WindowSizes.compact:
+              h = "Compact";
+            case WindowSizes.medium:
+              h = "Medium";
+            case WindowSizes.expanded:
+              h = "Expanded";
+          }
 
-            return Text("W: $w\nH: $h");
-          },
-        ),
+          return Text("W: $w\nH: $h");
+        },
       ),
     );
   }