+++ example/lib/main.dart
@@ -3,6 +3,7 @@ import 'package:flutter/material.dart';
 import 'package:responsive_ui/responsive_ux.dart';
 
 import 'builder_screen.dart';
+import 'listener_screen.dart';
 
 void main() {
   runApp(const ResponsiveUi());
@@ -19,7 +20,9 @@ class ResponsiveUi extends StatelessWidget {
           data: WindowSizeConfigurationData(
           sizes: [WindowSizes.compact]
         ),*/
-        child: BuilderScreen(),
+        child:
+          BuilderScreen(),
+          //ListenerScreen(),
       ),
     );
   }