+++ lib/src/ui/window_size_configuration.dart
@@ -1,5 +1,6 @@
 import 'package:flutter/widgets.dart';
 
+import '../data/window_size.dart';
 import '../data/window_size_configuration_data.dart';
 
 /// Should be inserted high in the Widget tree, e.g. beneath the [WidgetsApp]
@@ -28,12 +29,26 @@ class WindowSizeConfiguration extends StatelessWidget {
   static WindowSizeConfigurationData of(BuildContext context) {
     return context.dependOnInheritedWidgetOfExactType<_InheritedConfiguration>()?.data ?? _fallback;
   }
+
+  static (WindowSize, WindowSize)? windowSizeMaybeOf(BuildContext context) {
+    WindowSizeConfigurationData? config = maybeOf(context);
+
+    if (config == null) return null;
+
+    return config.getWindowSize(MediaQuery.sizeOf(context));
+  }
+
+  static (WindowSize, WindowSize) windowSizeOf(BuildContext context) {
+    WindowSizeConfigurationData config = of(context);
+
+    return config.getWindowSize(MediaQuery.sizeOf(context));
+  }
 }
 
 class _InheritedConfiguration extends InheritedWidget {
   final WindowSizeConfigurationData data;
 
-  const _InheritedConfiguration({super.key, required this.data, required super.child});
+  const _InheritedConfiguration({required this.data, required super.child});
 
   @override
   bool updateShouldNotify(covariant _InheritedConfiguration oldWidget) {